pyproject.toml
[project]
name = "resumeoptimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.117.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.25",
]
